tweaks and user input
